requirements.txt
tensorflow-gpu
opencv-python
numpy
tqdm
